chore: move keras_reader into keras module to avoid requiring tensorflow to be installed when using onnx functions
