+++ src/omlt/io/__init__.py
@@ -1,2 +1,1 @@
 from omlt.io.onnx import load_onnx_neural_network, write_onnx_model_with_bounds, load_onnx_neural_network_with_bounds
-from omlt.io.keras_reader import load_keras_sequential
