+++ src/omlt/io/keras/__init__.py
@@ -1,0 +1,1 @@
+from omlt.io.keras.keras_reader import load_keras_sequential
