+++ tests/neuralnet/test_keras.py
@@ -4,7 +4,7 @@ import pyomo.environ as pyo
 import numpy as np
 
 from omlt.block import OmltBlock
-from omlt.io.keras_reader import load_keras_sequential
+from omlt.io.keras import load_keras_sequential
 from omlt.neuralnet import (FullSpaceNNFormulation, ReducedSpaceNNFormulation)
 from omlt.neuralnet.activations import ComplementarityReLUActivation
 from omlt.scaling import OffsetScaling
